requirements.txt
git+https://github.com/Kewton/jsonDB
streamlit
python-dotenv
openai
